employeeProject/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.1.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
packaging==21.3
protobuf==3.20.0
PyMySQL==1.0.2
pyparsing==3.0.8
pytz==2022.1
sqlparse==0.4.2
wrapt==1.14.0
